requirements.txt
pyserial
influxdb
flask